Update setup.py to include IPython dep
The package expects the IPython library to be available but does not include in the dependency list under `setup.py`'s 'install_requires=[...]` list of string literals. This commit adds the dependency. 

Prior to this the work around needed is to-in a shell-in the python env run `pip install IPython`-after running `pip intall bertviz` and then do your stuffs.